# perf: manipulate arguments instead of creating new array