This is something that `util.promisify` does and I don't that has no harm.

Before:
```
util.promisify x 22,101,833 ops/sec ±0.60% (94 runs sampled)
universalify.fromCallback x 17,859,701 ops/sec ±0.48% (98 runs sampled)
universalify.fromPromise x 12,234,774 ops/sec ±1.19% (79 runs sampled)
```

Now:
```
util.promisify x 21,313,745 ops/sec ±0.93% (94 runs sampled)
universalify.fromCallback x 21,528,063 ops/sec ±0.76% (96 runs sampled)
universalify.fromPromise x 16,615,574 ops/sec ±0.83% (82 runs sampled)
```

<details>
<summary>benchmark</summary>

```js
import benchmark from 'benchmark';
import { promisify } from 'node:util';
import { fromCallback, fromPromise } from 'universalify';

const suite = benchmark.Suite();
const callbackTest = (test, fn) => {
  fn(null, test);
}
const promiseTest = async (test) => {
  return test;
}

const utilPromisifyTest = promisify(callbackTest);
const universalifyTest = fromCallback(callbackTest);

const universalifyCallback = fromPromise(promiseTest);

suite.add('util.promisify', () => {
  utilPromisifyTest('hello');
}).add('universalify.fromCallback', () => {
  universalifyTest('hello');
}).add('universalify.fromPromise', (deferred) => {
  universalifyCallback('hello', (t) => {
    deferred.resolve();
  });
}, { defer: true })
  .on('cycle', event => {
    console.log(event.target.toString())
  })
  .run({ async: false });
```